ST fw update for band selection at runtime
Command "+BAND" added to the predefined set of AT commands. It allows to select the frequency band at runtime.
The initial band is set to the European one. If the user selects a different band, the Lora state machine is
re-initialized and the desired band of frequency is enabled.
In the compiler builIn options, the symbols of the frequency bands of all the regions are set to False.

An interactive test is available here: https://github.com/giulcioffi/tests_MKRWAN/blob/master/band_test.ino
It continuously asks the user to select a new region. After the user inserts one, the value of the frequency
on the RX channel is printed. In case this value is different from the one that should be set for that region,
an error message is displayed.